fix(back): Add basic auth to save_metric endpoint too
